requirements.txt
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
python-dotenv==0.19.0
regex==2021.9.30
tomli==1.2.1
typing-extensions==3.10.0.2
Werkzeug==2.0.1
